Link the full engine into a macOS arm64 executable
Resolve the remaining ~810 undefined symbols at the final Game link by
compiling the real source files that were missing from the module
CMakeLists, and stubbing only the genuinely unportable leaves. The whole
engine now links: build/Game/Game (Mach-O arm64), 0 undefined / 0
duplicate symbols.

Real files pulled into the build (with clang-vs-MSVC fixes):
- Game: Actions, Conditions, GameOptions, IniFile, MissionDescription,
  StreamCommand, Universe, TriggerExport, ...
- Water (un-gated), Terra VMAP/LAND, XUtil XBUFFER/XSTREAM/XUTIL/XRnd,
  Physics Crash/RigidBodyNode, Util DebugUtil (portable debug-draw).

Stub leaves (gated non-WIN32, Track-B replacements later):
- Render D3D9 backend (RenderStub) + FreeType + sVideoWrite
- Sound (SoundStub), Network PNetCenter (NetCenterStub)
- Win32/editor toolkit: kdw widgets, ConsoleWindow, DebugSymbolManager,
  KeyboardDriver, joystick/DirectInput, Profiler, FormulaString,
  TriggerEditor, VideoMemoryInformation, SystemUtil GUI dialogs
- x86 cpuid, licensing (VerifyCDKey), main()->WinMain entry

Portable platform shims (Platform/WindowsAPI.h): GlobalAlloc/OpenFile,
templated ReadFile/WriteFile, itoa family; expanded fake d3d9.h.

Cross-platform correctness fixes:
- EnumDescriptor: macros now emit getEnumDescriptor<T> explicit
  specializations (call sites reference the template, not the overload).
- xm_inline -> inline off-Windows (__forceinline lacks ODR linkage on
  clang, causing duplicate math symbols at -O0).
- __declspec(selectany) globals (logicRnd, logMsgCenter) -> C++17 inline
  variables (__declspec is stripped off-Windows).
- vmap.inl explicit specializations marked inline; FileImage saveFileSmart
  renamed local to avoid an extern-promotion clash.

+++ Game/Conditions.cpp
@@ -439,7 +439,7 @@ bool ConditionCheckSurfaceProducedZone::check(UnitActing* unit) const
 
 	ScanGroundDamagedLineOp line_op;
 	ScanningShape shape;
-	shape.setCircle(round(unit->attr().producedPlacementZoneRadius) >> kmGrid);
+	shape.setCircle(int(round(unit->attr().producedPlacementZoneRadius)) >> kmGrid);
 	ScanningShape::const_iterator i;
 	int y = shape.rect().y + (unit->position2D().yi() >> kmGrid);
 	int xCenter = unit->position2D().xi() >> kmGrid;
@@ -623,7 +623,7 @@ bool ConditionObjectPercentOwner::check() const
 	}
 
 	totalMax = max(total, totalMax);
-	return compare(total ? round(part * 100.f / totalMax) : 0, percent, compareOperator); 
+	return compare(total ? int(round(part * 100.f / totalMax)) : 0, percent, compareOperator);
 }
 
 void ConditionCreateObject::serialize(Archive& ar)
